include altered orca runtime
In order to achieve the classic windows solitaire effect on game win, I
needed to disable clearing the frame buffer every frame. Currently orca
does not expose this functionality, so I had to hack it in myself. I am
including the compiled library liborca.a that has my changes to the
runtime.